storage efficiency, goerli migration

+++ contracts/Voting_Booth.sol
@@ -34,13 +34,13 @@ contract VotingBooth is VotingStorage {
     }
 
     //returns catId for new category given category name and candidates array
-    function setCategory(bytes32 _category, bytes32[] memory _candidates) external onlyOwner returns (uint64){
+    function setCategory(bytes32 _category, bytes32[] memory _candidates) external onlyOwner returns (uint128){
 
         //ensures number of candidates remains within amount allowable by data type for canId (uint8)
         require(_candidates.length <256, "There are too many candidates.");
 
         //grab current catId number
-        uint64 currentCatId = _catCounter;
+        uint128 currentCatId = _catCounter;
         _catCounter++;
 
         //create new categories array entry and populate category struct data
@@ -58,7 +58,7 @@ contract VotingBooth is VotingStorage {
     }
 
     //allows owner to open a previously set category given catId
-    function openCategory(uint64 _catId) external onlyOwner categoryClosed(_catId) {
+    function openCategory(uint256 _catId) external onlyOwner categoryClosed(_catId) {
 
         //only allow categories to be opened once
         require(_openedOnce[_catId] == false, "This category cannot be opened again");
@@ -69,7 +69,7 @@ contract VotingBooth is VotingStorage {
     }
 
     //allows owner to close an open category and determine winner given catId
-    function closeCategory(uint64 _catId) external onlyOwner categoryOpen(_catId) returns(bytes32) {
+    function closeCategory(uint256 _catId) external onlyOwner categoryOpen(_catId) returns(bytes32) {
 
         categories[_catId].open = false;
 
@@ -82,7 +82,7 @@ contract VotingBooth is VotingStorage {
     }
 
     //allows user to cast vote given category Id and candidate Id
-    function castVote(uint64 _catId, uint8 _canId) external registered whenNotPaused categoryOpen(_catId) {
+    function castVote(uint256 _catId, uint8 _canId) external registered whenNotPaused categoryOpen(_catId) {
 
         //prevent voter from voting for another candidate in the same category
         require(_boolVoter[msg.sender][_catId] == false, "You have already voted in this category");
@@ -98,7 +98,7 @@ contract VotingBooth is VotingStorage {
 // Getter functions ------------------------------------------------------------------------------------------------------------------------------------------------------
 
     //returns total number of category votes given catId
-    function getCategoryVotes(uint64 _catId) public view returns(uint) {
+    function getCategoryVotes(uint256 _catId) public view returns(uint) {
 
         //loop through cadidates and sum candidate votes
         uint categoryVotes = 0;
@@ -110,7 +110,7 @@ contract VotingBooth is VotingStorage {
     }
 
     //returns category winner given catId
-    function getCategoryWinner(uint64 _catId) internal view returns(bytes32) {
+    function getCategoryWinner(uint256 _catId) internal view returns(bytes32) {
 
         bytes32 categoryWinner = "no winner";
         uint categoryWinnerVotes = 0;
@@ -130,22 +130,22 @@ contract VotingBooth is VotingStorage {
     }
 
     //returns category name given category Id
-    function getCategoryName(uint64 _catId) external view returns(bytes32) {
+    function getCategoryName(uint256 _catId) external view returns(bytes32) {
         return categories[_catId].name;
     }
 
     //returns bool value showing if category open is true or false given category Id
-    function getCategoryOpen(uint64 _catId) external view returns(bool) {
+    function getCategoryOpen(uint256 _catId) external view returns(bool) {
         return categories[_catId].open;
     }
 
     //returns candidate name given both category and candidate Ids
-    function getCandidate(uint64 _catId, uint8 _canId) external view returns(bytes32) {
+    function getCandidate(uint256 _catId, uint8 _canId) external view returns(bytes32) {
         return categories[_catId].candidates[_canId].name;
     }
     
     //returns candidate total votes given category and candidate Ids as long as category is closed
-    function getCandidateVotes(uint64 _catId, uint8 _canId) external view categoryClosed(_catId) returns(uint) {
+    function getCandidateVotes(uint256 _catId, uint8 _canId) external view categoryClosed(_catId) returns(uint) {
         return categories[_catId].candidates[_canId].numCanVotes;
     }
